Remove the redundant EncodeMetaBlockHeader() function.
Use Store{Compressed,Uncompressed}MetaBlockHeader() instead.

+++ enc/brotli_bit_stream.cc
@@ -55,7 +55,7 @@ void StoreVarLenUint8(int n, int* storage_ix, uint8_t* storage) {
 }
 
 bool StoreCompressedMetaBlockHeader(bool final_block,
-                                    int length,
+                                    size_t length,
                                     int* storage_ix,
                                     uint8_t* storage) {
   // Write ISLAST bit.
@@ -90,7 +90,7 @@ bool StoreCompressedMetaBlockHeader(bool final_block,
   return true;
 }
 
-bool StoreUncompressedMetaBlockHeader(int length,
+bool StoreUncompressedMetaBlockHeader(size_t length,
                                       int* storage_ix,
                                       uint8_t* storage) {
   // Write ISLAST bit. Uncompressed block cannot be the last one, so set to 0.
